Move function to other file
