+++ index.js
@@ -4,11 +4,11 @@ import { ArticleManager } from './logic/article-manager.js';
 import { validateDuplicatedOpeningHeading, validateHeadingsNesting, validateMaxLength, validateNumberedHeadings, validateOpeningHeadingLevel, validateSpacing, validateTitleCase } from './validations/headings.js'
 import { validateMetaData } from './validations/metadata.js';
 import { validateRules } from './validations/rules.js';
-import { validateImageDescriptions, validateImagePaths, validateReferencedAssets, validateSVGFiles, validateAssetsFolderName } from './validations/assets.js';
+import { validateImageDescriptions, validateImagePaths, validateReferencedAssets, validateSVGFiles } from './validations/assets.js';
 import { validateSyntaxSpecifiers } from './validations/code-blocks.js';
 import { validateNestedLists } from './validations/lists.js';
 import { validateBrokenLinks } from './validations/links.js';
 import { ConfigManager } from './logic/config-manager.js';
-import { validateFolderName } from './validations/naming.js';
+import { validateFolderName, validateAssetsFolderName } from './validations/naming.js';
 
 export { Validator, ArticleManager, validateDuplicatedOpeningHeading, validateHeadingsNesting, validateMaxLength, validateNumberedHeadings, validateOpeningHeadingLevel, validateSpacing, validateTitleCase, validateMetaData, validateRules, validateImageDescriptions, validateImagePaths, validateReferencedAssets, validateSVGFiles, validateAssetsFolderName, validateSyntaxSpecifiers, validateNestedLists, validateBrokenLinks, ConfigManager, validateFolderName }
