+++ validations/naming.js
@@ -27,4 +27,26 @@ function validateFolderName(article) {
     return errorsOccurred;
 }
 
-export { validateFolderName }
+
+/**
+ * Checks if the article uses one of the allowed assets folder name.
+ * @param {Article} article 
+ * @param {String} expectedFolderName the expected assets folder name
+ * @returns an array of ValidationIssue objects for the found issues.
+ */
+ function validateAssetsFolderName(article, expectedFolderName){
+
+    if(article.assets.length > 0 && article.assetsFolder === null){
+        const errorMessage = "Multiple asset directories used";
+        return new ValidationIssue(errorMessage, article.contentFilePath);
+    }
+
+    if(article.assetsFolder !== expectedFolderName){
+        const errorMessage = "Unexpected or deprecated assets directory used";
+        return new ValidationIssue(errorMessage, article.contentFilePath,  ValidationIssue.Type.WARNING);
+    }     
+
+    return [];
+}
+
+export { validateFolderName, validateAssetsFolderName }
